Olympia: Playground updates

+++ start.sh
@@ -2,6 +2,7 @@
 set -e
 
 # Run a complete joystream development network on your machine using docker
+export JOYSTREAM_NODE_TAG=${JOYSTREAM_NODE_TAG:=$(./scripts/runtime-code-shasum.sh)}
 
 INIT_CHAIN_SCENARIO=${INIT_CHAIN_SCENARIO:=setupNewChain}
 
@@ -25,13 +26,19 @@ fi
 docker-compose up -d joystream-node
 
 ## Init the chain with some state
-export SKIP_MOCK_CONTENT=true
-export SKIP_QUERY_NODE_CHECKS=true
-HOST_IP=$(tests/network-tests/get-host-ip.sh)
-export COLOSSUS_1_URL=${COLOSSUS_1_URL:="http://${HOST_IP}:3333"}
-export COLOSSUS_1_TRANSACTOR_KEY=$(docker run --rm --pull=always docker.io/parity/subkey:2.0.1 inspect ${COLOSSUS_1_TRANSACTOR_URI} --output-type json | jq .ss58Address -r)
-export DISTRIBUTOR_1_URL=${DISTRIBUTOR_1_URL:="http://${HOST_IP}:3334"}
-./tests/integration-tests/run-test-scenario.sh ${INIT_CHAIN_SCENARIO}
+if [[ $SKIP_CHAIN_SETUP != 'true' ]]; then
+  set -a
+  . ./.env
+  set +a
+
+  export SKIP_MOCK_CONTENT=true
+  export SKIP_QUERY_NODE_CHECKS=true
+  HOST_IP=$(tests/network-tests/get-host-ip.sh)
+  export COLOSSUS_1_URL=${COLOSSUS_1_URL:="http://${HOST_IP}:3333"}
+  export COLOSSUS_1_TRANSACTOR_KEY=$(docker run --rm --pull=always docker.io/parity/subkey:2.0.1 inspect ${COLOSSUS_1_TRANSACTOR_URI} --output-type json | jq .ss58Address -r)
+  export DISTRIBUTOR_1_URL=${DISTRIBUTOR_1_URL:="http://${HOST_IP}:3334"}
+  ./tests/integration-tests/run-test-scenario.sh ${INIT_CHAIN_SCENARIO}
+fi
 
 ## Member faucet
 docker-compose up -d faucet
